[examples] Cleanup after `print` now uses templates
Fixed-Integer width print templates were introduced in 962b60c

+++ examples/example-i2c-bh1750/src/app_main.cpp
@@ -89,7 +89,7 @@ void ReadLux() {
     float light = bh.readLightLevel();
     // Switch off the info LED if light is low else on
     digitalWrite(PIN_INFO, (light < 50));
-    serial.println("Lux: ", (int)light);
+    serial.println("Lux: ", light);
   }
 }
